Sprint 3

+++ database/migrations/2020_12_20_193556_modify_group_column_of_admin_table.php
@@ -15,10 +15,13 @@ class ModifyGroupColumnOfAdminTable extends Migration
     {
         Schema::table('groups', function (Blueprint $table) {
             //
-            $table->dropForeign(['admin']);
-            $table->dropColumn('admin');
-            $table->unsignedBigInteger('admin');
-            $table->foreign('admin')->references('id')->on('users')->onDelete('cascade');
+            if(!Schema::hasColumn('groups','admin')){
+                $table->dropForeign(['admin']);
+                $table->dropColumn('admin');
+                $table->unsignedBigInteger('admin');
+                $table->foreign('admin')->references('id')->on('users')->onDelete('cascade');
+
+            }
         });
     }
 
@@ -31,7 +34,7 @@ class ModifyGroupColumnOfAdminTable extends Migration
     {
         Schema::table('groups', function (Blueprint $table) {
             //
-
+            $table->dropColumn('admin');
         });
     }
 }
